+++ Source/tests/ClipMate.Tests.Integration/Services/ClipPersistenceTests.cs
@@ -3,7 +3,6 @@ using ClipMate.Core.Services;
 using ClipMate.Data;
 using ClipMate.Data.Repositories;
 using ClipMate.Data.Services;
-using ClipMate.Platform;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Moq;